mplayer_server: change the type of callback_t
we need to access the opcode of the callbacks from outside of the server
(i.e. in the future clients).

the type callback_t is now a pointer to function and can be accessed
using callbacks_g[opcode] where the opcodes are enumerated in request.h

a new file named pub_callbacks.h has been added.
it will be used later by the client and must not be server dependant

closes #1